Refactor Project Name & Packages
Originally had project name set as "EmptyActivity", now it is "GymPlanner" to be seamless with our actual name chosen. Same was done for all packages & in google_services.json
