+++ app/google-services.json
@@ -9,7 +9,7 @@
       "client_info": {
         "mobilesdk_app_id": "1:474565980607:android:7b4463e49ccfc587d2310a",
         "android_client_info": {
-          "package_name": "com.example.workouttracker"
+          "package_name": "com.example.gymplanner"
         }
       },
       "oauth_client": [],
