+++ app/src/main/java/com/example/gymplanner/exerciselist/ExerciseList.kt
@@ -10,7 +10,7 @@
     ExperimentalMaterial3Api::class, ExperimentalMaterial3Api::class
 )
 
-package com.example.workouttracker.exerciselist
+package com.example.gymplanner.exerciselist
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
@@ -41,7 +41,7 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import com.example.workouttracker.ui.theme.WorkoutTrackerTheme
+import com.example.gymplanner.ui.theme.WorkoutTrackerTheme
 import androidx.compose.foundation.lazy.items
 import androidx.compose.material3.BottomAppBar
 import androidx.compose.material3.Scaffold
