+++ app/src/main/java/com/example/gymplanner/ui/theme/Type.kt
@@ -1,4 +1,4 @@
-package com.example.workouttracker.ui.theme
+package com.example.gymplanner.ui.theme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
